+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [NEXT] - YYYY-MM-DD
+- Made compatible with ItemPhysic mod
+
 ## [0.0.3] - 2019-12-07
 - Modularized bounding boxes (should have no visible change)
 - Made item insertion and inventory updating server-side only